requirements.txt
streamlit==1.5.1
pandas==1.3.5
numpy==1.19.5
base58==2.1.1
matplotlib==3.4.3
pillow==8.4.0
scikit-learn==1.0.1
shap==0.40.0
rdkit==2021.9.4
